os::time returns long; seed is long

+++ src/rbc/rnd/imp.h
@@ -1,7 +1,7 @@
 namespace rbc { namespace rnd {
 struct D;
 enum {ENV = -2, TIME = -1}; /* special seeds */
-void ini(D**, int n, int seed);
+void ini(D**, int n, long seed);
 void fin(D*);
 void gen(D*, int n);
 float get_hst(const D*, int i);
