+++ src/utils/os.h
@@ -1,4 +1,4 @@
 namespace os { /* [o]perating [s]ystems commands */
 void mkdir(const char* path);
-int time();
+long time();
 }
